[TITLE] update runtime go111
[CHANGES]
- split functions
- remove AppEngineTokenSource and use DefaultTokenSource
