+++ controller/error.go
@@ -1,0 +1,18 @@
+package controller
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func errorHandler(w http.ResponseWriter, r *http.Request, status int) {
+	w.WriteHeader(status)
+	switch status {
+	case http.StatusNotFound:
+		fmt.Fprint(w, "404 Not Found")
+	case http.StatusMethodNotAllowed:
+		fmt.Fprint(w, "405 Method Not Allow")
+	default:
+		fmt.Fprint(w, "Bad Request")
+	}
+}
