+++ controller/helper.go
@@ -1,0 +1,17 @@
+package controller
+
+import (
+	"context"
+	"os"
+
+	"google.golang.org/appengine"
+)
+
+func getProjectID(ctx context.Context) (projectID string) {
+	if appengine.AppID(ctx) == "None" {
+		projectID = os.Getenv("PROJECT_ID")
+	} else {
+		projectID = appengine.AppID(ctx)
+	}
+	return
+}
