+++ go.mod
@@ -1,0 +1,8 @@
+module github.com/CloudMile/gae_snapshot_gce_disk
+
+require (
+	cloud.google.com/go v0.34.0 // indirect
+	golang.org/x/oauth2 v0.0.0-20181203162652-d668ce993890
+	google.golang.org/api v0.0.0-20181212003324-40e757e92c52
+	google.golang.org/appengine v1.3.0
+)
